refactor: simplify nutrition optimizer frontend
