+++ src/diet/nutrition_optimizer/static/style.css
