+++ src/diet/nutrition_optimizer/templates/nutrition_optimizer/_constraints.html
@@ -31,7 +31,12 @@
                 </select>
             </td>
             <td>
-                <input type="number" class="form-control" name="constraint-value" />
+                <input type="number"
+                       class="form-control"
+                       name="constraint-value"
+                       min="0"
+                       step="any"
+                       required />
             </td>
             <td class="no-action"></td>
         </tr>
@@ -64,7 +69,12 @@
             </select>
         </td>
         <td>
-            <input type="number" class="form-control" name="constraint-value" />
+            <input type="number"
+                   class="form-control"
+                   name="constraint-value"
+                   min="0"
+                   step="any"
+                   required />
         </td>
         <td>
             <button type="button" class="btn btn-outline-danger">Remove</button>
